Split `cli` module and integrate `config` subcommand permanently
