--- src/elva/commands/config/__init__.py
@@ -1,7 +1,0 @@
-"""
-Configure TLS options.
-"""
-
-from .cli import cli
-
-__all__ = [cli]
