+++ tests/test_server.py
@@ -1,7 +1,7 @@
 import sqlite3
+import tomllib
 import uuid
 from http import HTTPStatus
-import tomllib
 
 import anyio
 import pytest
